Fix two more CodeQL integer-conversion alerts in procscan
Same root pattern as the PPID fix already merged (PR #3): parsing into
an architecture-dependent int, then narrowing with an unchecked cast,
instead of parsing directly into the destination width.

- Snapshot: strconv.Atoi(entry.Name()) (int) followed by uint32(pid) at
  the call into scanPID. Changed to strconv.ParseUint(..., 10, 32)
  directly, and threaded uint32 through Snapshot/scanPID's signatures
  instead of int. (alert #7)

- scanPID: time.Duration(startTicks) casts a raw uint64 straight to
  int64 (time.Duration's underlying type) with no bounds check. An
  implausibly large startTicks would wrap to a negative duration
  silently. Extracted the computation into ticksToDuration, which
  bounds-checks against math.MaxInt64 first. Confirmed directly: without
  the check, ticksToDuration(math.MaxUint64) returns -10ms instead of
  failing. (alert #6)

Both come with tests (TestTicksToDuration exercises the new bounds
check in isolation; TestSnapshotFindsSelfAndSkipsSelfPID updated for
the uint32 signature change, replacing the old -1 sentinel with
^uint32(0)). Verified live: procscan.Snapshot still seeds correctly
against the real /proc during an actual `provctl watch` run (471
processes). go build/vet/test/-race/gofmt all clean.

+++ internal/cli/watch.go
@@ -46,7 +46,7 @@ func runWatch(ctx context.Context, args []string) error {
 	// Seed the store with processes that already exist, so events from
 	// long-lived processes (shells, browsers) can be attributed to a named
 	// process instead of a bare pid we never saw start.
-	if snapshot, err := procscan.Snapshot(os.Getpid()); err != nil {
+	if snapshot, err := procscan.Snapshot(uint32(os.Getpid())); err != nil {
 		fmt.Fprintln(os.Stderr, "provctl: /proc snapshot failed, pre-existing processes will show as bare pids:", err)
 	} else if err := st.Seed(snapshot); err != nil {
 		return fmt.Errorf("cli: seed store from /proc: %w", err)
